2023 - day 5, part 1 - If You Give A Seed A Fertilizer
The puzzle input is an almanac describing the seeds that need to be planted and mappings from seed
to location with various intermediate steps. The mappings are defined by destination, source, and
range numbers describing how to convert from the source number to the destination number. The puzzle
answer is the lowest location number of any of the seeds which corresponds to the closest location.