2023 - day 6, part 1 - Wait For It
The puzzle input is a few races with the total allotted time in milliseconds and the record distance
in millimeters. The race is done with toy boats that gain speed for each millisecond a button is pushed
down. The puzzle answer is for each race finding the ways you can beat the record distance and multiplying
them together to get the "margin of error"