2023 - day 6, part 2, Wait For It
The puzzle changed to now ignore the whitespace that was delimiting each race. This means that the
input is now one giant race with one overall time and one overall distance. This seemed to punish
anyone who brute forced part 1 but the binary search solution I implemented was able to get the new
puzzle answer of how many ways to there are to beat this new big race. The only change was to parsing the input.